+++ GameEngine1/Interfaces/IHealth.cs
@@ -6,7 +6,7 @@ namespace GameEngine1.Interfaces
 {
     public interface IHealth
     {
-        public int Health { get; set; }
+        public float Health { get; set; }
         public int MaxHealth { get; set; }
         public bool Hit { get; set; }
     }
